refactor: extract formatTimeAgo to shared utils
Move duplicated formatTimeAgo from panels.tsx and
RecentSourcesModal.tsx to frontend/src/lib/utils.ts.
