+++ frontend/src/components/RecentSourcesModal.tsx
@@ -10,7 +10,7 @@
  */
 
 import * as React from 'react';
-import { cn } from '@/lib/utils';
+import { cn, formatTimeAgo } from '@/lib/utils';
 import {
   getSavedDataSources,
   reconnectSavedSource,
@@ -27,18 +27,6 @@ interface RecentSourcesModalProps {
 }
 
 // --- Helpers ---
-
-function formatTimeAgo(dateStr: string): string {
-  const diff = Date.now() - new Date(dateStr).getTime();
-  const minutes = Math.floor(diff / 60000);
-  const hours = Math.floor(diff / 3600000);
-  const days = Math.floor(diff / 86400000);
-
-  if (days > 0) return `${days}d ago`;
-  if (hours > 0) return `${hours}h ago`;
-  if (minutes > 0) return `${minutes}m ago`;
-  return 'Just now';
-}
 
 function sourceIcon(type: string) {
   if (type === 'database') return <DatabaseIcon className="w-4 h-4 text-amber-400" />;
